Ignore values AND'd with 0 for more than 1b case
Since &ing a value with 0 is a common idiom, this change makes the graph
pruning a little less conservative than it needs to be.
In RVMiniDatapath, this decreased the number of used wires from 96135000
to 961305000.

Similar optimizations can't really be made for other binary operations
as I don't know of other similar idioms (except XORing a value with
itself).